Moved tests to folders, and added salt to app.config
