+++ Tests/FileCache/Controller_FileCacheShould.cs
@@ -3,7 +3,7 @@ using NUnit.Framework;
 using System;
 using System.IO;
 
-namespace Tests
+namespace Tests.FileCache
 {
     [TestFixture]
     public class Controller_FileCacheShould
@@ -13,7 +13,7 @@ namespace Tests
         [SetUp]
         public void SetUp()
         {
-            Path = FileCache.Instance.GetFile("songs/untrago.mp3");
+            Path = Controller.FileCache.Instance.GetFile("songs/untrago.mp3");
         }
 
         [Test]
@@ -25,7 +25,7 @@ namespace Tests
         [Test]
         public void ClearCache_Path_NotExists()
         {
-            FileCache.Instance.ClearCache(TimeSpan.FromSeconds(-1));
+            Controller.FileCache.Instance.ClearCache(TimeSpan.FromSeconds(-1));
             Assert.IsFalse(File.Exists(Path));
         }
     }
