+++ Tests/Playlists/ControllorDeactivate-DeletePlaylist.cs
@@ -7,7 +7,7 @@ using Model.DbModels;
 using NUnit.Framework;
 using Assert = NUnit.Framework.Assert;
 
-namespace Tests
+namespace Tests.Playlists
 {
     [TestFixture]
     public class ControllorDeactivate_DeletePlaylist
